Fix everything up
