--- app/controllers/api/auth_tokens_controller.rb
@@ -1,11 +1,0 @@
-module API
-  class AuthTokensController < BaseController
-    def reset
-      current_user.create_auth_token!
-
-      render json: {
-        auth_token: current_user.auth_token
-      }
-    end
-  end
-end
