+++ app/controllers/api/settings/auth_tokens_controller.rb
@@ -1,0 +1,13 @@
+module API
+  module Settings
+    class AuthTokensController < BaseController
+      def reset
+        current_user.create_auth_token!
+
+        render json: {
+          auth_token: current_user.auth_token
+        }
+      end
+    end
+  end
+end
