--- app/controllers/settings/api_controller.rb
@@ -1,5 +1,0 @@
-module Settings
-  class APIController < ApplicationController
-    def show; end
-  end
-end
