+++ app/views/settings/api_cli.html.haml
@@ -1,6 +1,6 @@
 #page-settings-api.page-settings
   .lg-container
-    = render ViewComponents::SettingsNav.new(:api)
+    = render ViewComponents::SettingsNav.new(:api_cli)
 
     %article
       %h1 API &amp; CLI Settings
