+++ app/views/settings/communication_preferences.html.haml
@@ -3,7 +3,7 @@
     = render ViewComponents::SettingsNav.new(:communication)
 
     %article
-      %h1 Communication
+      %h1 Communication Preferences
 
       %section
         = render ReactComponents::Settings::CommunicationPreferencesForm.new
